chore: change file name Walls.hs to Map.hs
